cmd/dictgen: Added option to prevent the footer from showing on certain locales (fixes #5)
Some locales, like French, add a footer showing the source of the dictionary
due to licencing restrictions, which aren't applicable to our custom ones. This
option adds the HTML,

  <span class="end"><style>.end,.end+*{display: none !important;}</style></span>

to the end of each entry's generated HTML, which prevents the footer, which is
added as plain text in another block element, from being displayed.

This option should not be used unless necessary (i.e. replacing one of those
dictionaries), as it bloats the HTML code and is a dirty hack to avoid needing
to patch the firmware.